requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=3.0.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
gspread>=6.0.0
google-auth>=2.0.0
twilio>=9.0.0
pytest>=8.0.0
httpx>=0.27.0
